split classes and build rspy package
