+++ unit-tests/algo/d2rgb/test-triggers.py
@@ -1,4 +1,5 @@
-import pyrealsense2 as rs, test, ac
+import pyrealsense2 as rs, test
+from rspy import test, ac
 
 # We set the environment variables to suit this test
 test.set_env_vars({"RS2_AC_DISABLE_CONDITIONS":"1",
