+++ unit-tests/func/ir/l500/test-frame-filter.py
@@ -3,8 +3,9 @@
 
 #test:device L500*
 
-import pyrealsense2 as rs, test
-from pyrsutil import timer
+import pyrealsense2 as rs
+from rspy.timer import Timer
+from rspy import test
 import time
 
 # The L515 device opens the IR stream with the depth stream even if the user did not ask for it (for improving the depth quality), 
@@ -14,7 +15,7 @@ import time
 #   2. User ask for depth + IR - make sure he gets depth + IR frames
 
 MAX_TIME_TO_WAIT_FOR_FRAMES = 10 # [sec]
-NUMBER_OF_FRAMES_BEFORE_CHECK = 100 
+NUMBER_OF_FRAMES_BEFORE_CHECK = 50 
 
 devices = test.find_devices_by_product_line_or_exit(rs.product_line.L500)
 device = devices[0]
@@ -37,7 +38,7 @@ def frames_counter(frame):
         global n_ir_frame
         n_ir_frame += 1
 
-wait_frames_timer = timer(MAX_TIME_TO_WAIT_FOR_FRAMES)
+wait_frames_timer = Timer(MAX_TIME_TO_WAIT_FOR_FRAMES)
 
 # Test Part 1
 test.start("Ask for depth only - make sure only depth frames arrive")
@@ -54,11 +55,14 @@ while (not wait_frames_timer.has_expired()
 if wait_frames_timer.has_expired():
     print(str(NUMBER_OF_FRAMES_BEFORE_CHECK) + " frames did not arrived at "+ str(MAX_TIME_TO_WAIT_FOR_FRAMES) + " seconds , abort...")
     test.fail()
+else:
+    test.check(n_depth_frame >= NUMBER_OF_FRAMES_BEFORE_CHECK)
+    test.check_equal(n_ir_frame, 0)
 
-test.check(n_depth_frame >= NUMBER_OF_FRAMES_BEFORE_CHECK)
-test.check_equal(n_ir_frame, 0)
 depth_sensor.stop()
 depth_sensor.close()
+
+time.sleep(1) # Allow time to ensure no more frame callbacks after stopping sensor
 
 test.finish()
 
@@ -74,16 +78,19 @@ wait_frames_timer.start()
 
 # we wait for first NUMBER_OF_FRAMES_BEFORE_CHECK frames OR MAX_TIME_TO_WAIT_FOR_FRAMES seconds
 while (not wait_frames_timer.has_expired() 
-    and n_depth_frame + n_ir_frame < NUMBER_OF_FRAMES_BEFORE_CHECK):
+    and (n_depth_frame == 0 or n_ir_frame == 0)):
     time.sleep(1)
 
 if wait_frames_timer.has_expired():
     print(str(NUMBER_OF_FRAMES_BEFORE_CHECK) + " frames did not arrived at "+ str(MAX_TIME_TO_WAIT_FOR_FRAMES) + " seconds , abort...")
     test.fail()
+else:
+    test.check(n_depth_frame != 0)
+    test.check(n_ir_frame != 0)
 
-test.check(n_depth_frame != 0)
-test.check(n_ir_frame != 0)
-test.check(n_ir_frame + n_depth_frame >= NUMBER_OF_FRAMES_BEFORE_CHECK)
+depth_sensor.stop()
+depth_sensor.close()
+
 test.finish()
 
 test.print_results_and_exit()
