+++ unit-tests/func/test-set-option.py
@@ -2,7 +2,8 @@
 # Copyright(c) 2020 Intel Corporation. All Rights Reserved.
 
 import platform
-import pyrealsense2 as rs, test
+import pyrealsense2 as rs
+from rspy import test
 import time
 
 dev = test.find_first_device_or_exit()
