+++ unit-tests/py/rspy/stopwatch.py
@@ -4,7 +4,7 @@ import time
 
 # Timer that counts forward in time (vs backwards in the 'timer' class)
 # It supply a basic stopwatch API, reset, get elapsed time..
-class stopwatch:
+class Stopwatch:
 
     _start = 0
 
@@ -22,27 +22,4 @@ class stopwatch:
     # Get stopwatch start time
     def get_start(self):
         return self._start
-
-# A timer counting backwards in time(vs forwards in the `stopwatch` class )
-# It supply basic timer API, start, has_expired..
-class timer:
-
-    _delta = 0
-    _sw = stopwatch()
-
-    def __init__(self, timeout):
-        self._delta = timeout
-
-    # Start timer
-    def start(self):
-        self._sw.reset()
-
-    # Check if timer time expired
-    def has_expired(self):
-        return self._sw.get_start() + self._delta <= time.perf_counter()
-
-    # Force time expiration
-    def set_expired(self):
-        self._sw.reset(time.perf_counter() - (self._delta + 0.00001))
-
